+++ tests/e2e/cases/async_await_while_condition/in.ts
@@ -148,6 +148,17 @@ async function chooseDirectAwaitUninitializedVarTryFinallyPrelude(prefix: string
     return await laterBodyValue(source);
 }
 
+async function chooseDirectAwaitUninitializedVarTryCatchPrelude(prefix: string): Promise<string> {
+    var source: any;
+    try {
+        throw "catch";
+    } catch (reason) {
+        void reason;
+        source = prefix + "-catch";
+    }
+    return await laterBodyValue(source);
+}
+
 class MutableUninitializedVarAwaitChooser {
     async choose(prefix: string): Promise<string> {
         var source: string;
@@ -3807,6 +3818,7 @@ chooseDirectAwaitUninitializedVarForOfPrelude("direct-uninitialized-var-for-of")
 chooseDirectAwaitUninitializedVarForInPrelude("direct-uninitialized-var-for-in").then((value) => console.log("await-direct-uninitialized-var-for-in", value));
 chooseDirectAwaitUninitializedVarSwitchPrelude("direct-uninitialized-var-switch").then((value) => console.log("await-direct-uninitialized-var-switch", value));
 chooseDirectAwaitUninitializedVarTryFinallyPrelude("direct-uninitialized-var-try-finally").then((value) => console.log("await-direct-uninitialized-var-try-finally", value));
+chooseDirectAwaitUninitializedVarTryCatchPrelude("direct-uninitialized-var-try-catch").then((value) => console.log("await-direct-uninitialized-var-try-catch", value));
 new MutableUninitializedVarAwaitChooser().choose("method-mutable-uninitialized-var").then((value) => console.log("await-method-mutable-uninitialized-var-prelude", value));
 new MutableUninitializedVarAwaitChooser().chooseThrow("method-mutable-uninitialized-var").catch((reason) => console.log("await-method-mutable-uninitialized-var-throw", reason));
 chooseMutableUninitializedVarAwaitValue("value-mutable-uninitialized-var").then((value) => console.log("await-value-mutable-uninitialized-var-prelude", value));